LayoutTests/storage/indexeddb/intversion-invalid-setversion-has-no-side-effects-expected.txt

Test that when setVersion is called on a db with an integer version, versionchange events don't get fired on other open dbs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-invalid-setversion-has-no-side-effects.html"
indexedDB.deleteDatabase(dbname)

deleteSuccess():
request = indexedDB.open(dbname, 2)

firstUpgradeNeededCallback():
gotFirstUpgradeNeededEvent = true
PASS event.oldVersion is 0
PASS event.newVersion is 2

firstSuccessEventHandler():
PASS gotFirstUpgradeNeededEvent is true
connection1 = event.target.result
connection1.onversionchange = unexpectedVersionChangeCallback
PASS String(connection1) is "[object IDBDatabase]"
request = indexedDB.open(dbname)

secondSuccessCallback():
db = event.target.result
request = db.setVersion('version')
request.onerror = setVersionError
request.onsuccess = unexpectedSuccessCallback
request.onblocked = unexpectedBlockedCallback

setVersionError():
Open up a new db to ensure that if connection1 were to get a versionchange event, it will be delivered before the test finishes
new_db = indexedDB.open(dbname + '2')
new_db.onsuccess = newDBSuccess

newDBSuccess():
PASS successfullyParsed is true

TEST COMPLETE
